Bot position fix
I was not able to increase the width of the container of chatbot so we shifted from bubble opening to button triggered opening. Modified JS and custom CSS such that I could center it on my Website as well as make changes into chatbot container.
Added a background as well for the container which gives an actual feel of travel support chatbot. I would like to remind you that its not just a support bot, It is your online ai travel companion.